dev: 3 new public method API for CleverForm
validate()  - Trigger CleverFormData form validation at runtime with out button on click.
reset()     - Reset the form, remove the error messages in the DOM and return the data object that are reset.
rawData()   - Return the unvalidated form fields data.

Changes in the modified files because of the implementation of the new 3 API.